Moved template function to impl and reverted other changes
